Mock metrics client implementation
