--- inference_perf/metrics/client.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Metrics_Client(ABC):
-    @abstractmethod
-    def __init__(self) -> None:
-        pass
